+++ src/main/java/com/fin/testtask/dto/AmountDTO.java
@@ -3,13 +3,13 @@ package com.fin.testtask.dto;
 import java.math.BigDecimal;
 
 public class AmountDTO {
-    private Double amount;
+    private BigDecimal amount;
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
 }
